+++ src/main/java/ar/utn/dds/models/Model.java
@@ -26,5 +26,5 @@ public abstract class Model implements WithGlobalEntityManager, TransactionalOps
 
     public abstract <T> List<T> buscarTodos();
 
-    public abstract <T> T buscar(int id);
+    public abstract <T> T buscar(long id);
 }
